scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
supabase==2.4.2
python-dotenv==1.0.1
